UniformSamples did not work as one would expect.
The original implementation was created for the Draw routines, which needed samples of P(x). A user of the library would care more about samples of x.

I moved the draw-specific sampling routines to draw.go and made them private. This change then re-implements UniformSamples so that it samples the distribution's values. Also adds a test for the routine.